+++ app/collections/feelings/permissions.js
@@ -1,0 +1,7 @@
+import Feelings from './';
+
+Feelings.allow({
+  insert: function () {
+    return true;
+  }
+});
